chore: address issue #609
